Fix infinite loop in circe -> http4s encoder conversion
Humanity can not be made equal by declarations on paper.
Unless the material conditions for equality exist, it is
worse than mockery to pronounce men equal. And unless
there is equality (and by equality I mean equal chances
for every one to make the most of himself) unless, I say,
these equal chances exist, freedom, either of thought,
speech, or action, is equally a mockery.

Voltairine de Cleyre — In Defense of Emma Goldman and the
Right of Expropriation (1893)